+++ src/clients/clientSignup.js
@@ -42,7 +42,7 @@ const ClientSignup = props => {
     return (
         <div>
             <form onSubmit={handleSubmit}>
-                <h4 className="clientSignUpTitle">Client Sign Up</h4>
+                <h3>Sign Up</h3>
                 <input
                     type="text"
                     placeholder="Full Name"
@@ -68,39 +68,21 @@ const ClientSignup = props => {
                     required
                 />
 
-                <br />
-                <Link to="/login" className="already">I already have a client account</Link>
-                <br />
-                <Button type="submit">
-                    {isFetching ? (< Spinner className="mx-2"
-
-
                 <button type="submit">
                     {/* {/* {!credentials && < Spinner
-
                         as="span"
                         animation="border"
                         size="sm"
                         role="status"
                         aria-hidden="true"
-
-                    />) : ""}
-
                     />) : ""} */}
-
 
                     Sign Up
                 </button>
                 <br />
-
-                <Link exact to="/"><Button>Home</Button></Link>
-
                 <Link to="/login">I already have a client account</Link>
                 <br></br>
                 <Link exact to="/"><button>Home</button></Link>
-
-
-
 
             </form>
         </div>
