+++ src/index.css
@@ -1,93 +1,13 @@
 body {
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
+    'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
+    sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  height: 100%;
-}
-
-.App {
-  background: linear-gradient(#0ae4eb, #46acc2, #75c0be, white);
-  height: 25rem;
-}
-
-a {
-  /* border: 2px solid white; */
-  color: white;
-  text-decoration: none;
-  font-size: 2em;
-}
-
-a:hover {
-  text-decoration: none;
-  color: white;
-}
-
-a:visited {
-  color: white;
-}
-
-.already {
-  font-size: 1.5em;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New', monospace;
-}
-
-form {
-  display: flex;
-  flex-flow: column nowrap;
-  align-items: center;
-}
-
-.cardclass input {
-  border: none;
-  width: 100%;
-}
-
-label {
-  margin-bottom: -12px;
-  text-align: left;
-  width: 200px;
-}
-
-.cardclass {
-  display: flex;
-  justify-content: center;
-  padding: 1rem;
-}
-
-h1, h2, h3, h4, h5 {
-  color: white;
-}
-
-.title {
-  font-size: 3em;
-}
-
-.homepagetitle {
-  font-size: 2em;
-  font-style: italic;
-  margin-top: -6em;
-}
-
-.homepage {
-  background-image: url(./images/Background\ Image.png);
-  background-size: contain;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-}
-
-.clientSignUpTitle {
-  font-size: 2.1em;
-  font-style: italic;
-}
-
-.instructorSignUpTitle {
-  font-style: italic;
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
 }
