--- src/instructorBackup.js
@@ -1,17 +1,0 @@
-// import React from "react";
-// import { Link } from "react-router-dom";
-
-// const InstructorClasses = () => {
-//     return (
-//         <div>
-//             <h4>
-//                 This is where instructor class view will be
-//         </h4>
-//             <Link exact to="/"><button>Home</button></Link>
-
-//         </div>
-//     )
-
-// }
-
-// export default InstructorClasses;
